Add react-search-autocomplete to repo
Made a tiny change to allow a onHideResults function (dead easy
to implement) which wasn't implemented on the original repo.
Maybe I'm a complete noob, but at this point in time, this seems
the best way to do this. Quite frankly, I can't be arsed to
fiddle about with making a new repo which is a fork of this repo
and then import that here. It's like 15 files, if you have a
problem with that, you can do one.