Add progress bar with levels to concept overlay
Also added the levels to the progress bar shown on the question modal.
Unsure whether it makes sense to have the progress bar in two places
once we move the modal to not cover the overlay. I think not and I should
perhaps reconfigure the dots to work. Although if there's an uncertain
number of questions in a set then it's confusing.