Always show concept name when moused over
Both make the moused-over concept name larger than previously
and make it show smaller when zoomed out. This means it'll always
show on the original map. This hasn't been tested on all possible
map sizes (enormous ones may still struggle, but much less so as
the max font size scales with the square root of 1/zoom)